requirements.txt
psutil>=5.9.0
colorama>=0.4.6
tabulate>=0.9.0
PyQt6>=6.4.0
wmi>=1.5.1
py-cpuinfo>=9.0.0
pywin32>=306
requests>=2.31.0